services/training/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pandas
numpy
scikit-learn
joblib
mlflow
